Python solution for meetup problem
exercism.io python track - meetup solution

includes more complicated (backup) as well as refarctored solution
